Update SBUS decoding algorithm
The software is updated to check for the Footer and Header byte within a
single evaluation.  With the previous design, it was identified that if
the last byte of data was 0x00 (i.e. same a Footer byte) and the state
machine was out of sync with the data stream, the state machine would not
re-sync with the data stream.  This was the case because the last data
byte (0x00) would be identified as the footer, the Footer (0x00) would not
be the header value (0x0F) causing the state machine to re-initialized and
the entire packet to be missed.